Move some methods to utils for tests
